Added the route for adding a wifi coupons to coupons collection

+++ routes/api.js
@@ -130,32 +130,54 @@ module.exports = function(router) {
         });
     });
 
+    // Route for adding a Wifi coupon to the coupons collections
+    router.post('/addWifiCoupon', function(req, res){
+        var coupId = req.body.coupId;
+        var coupPass = req.body.coupPass;
+        var coupon = new Coupon({
+            couponId: coupId,
+            couponPassword: coupPass
+        });
+        coupon.save(function(err){
+            if (err){
+                console.log(err);
+                res.json({success: false, message: "An error occurred"});
+            } else {
+                res.json({success: true, message: "Wifi coupon added successfully"});
+            }
+        });
+    });
+
     // Route for activating a wifi coupon for the user
     // The coupon will be in the user object like => wifi: <couponCode> <couponPassword>
     router.post('/activate', function(req, res){
         var qrcode = req.body.qrcode;
         Coupon.find({}).exec(function(err, coupons){
-            var coupon = coupons[0];
-            User.findOneAndUpdate({qrcode: qrcode}, {wifi: coupon.couponId + " " + coupon.couponPassword}).exec(function(err, outputUser){
-                if (err){
-                    console.log(err);
-                    res.json({success: false, message: "An error occurred"});
-                } else {
-                    if (!outputUser){
-                        res.json({success: false, message: "No such user exists"});
+            if (len(coupons) <= 0){
+                res.json({success: false, message: "No more wifi coupons available"});
+            } else {
+                var coupon = coupons[0];
+                User.findOneAndUpdate({qrcode: qrcode}, {wifi: coupon.couponId + " " + coupon.couponPassword}).exec(function(err, outputUser){
+                    if (err){
+                        console.log(err);
+                        res.json({success: false, message: "An error occurred"});
                     } else {
-                        // Removing that coupon from the coupons collection because that coupon is no longer availableto other users
-                        Coupon.findOneAndRemove({_id: coupon._id}).exec(function(err){
-                            if (err){
-                                console.log(err);
-                                res.json({success: false, message: "Error occurred while activating the coupon"});
-                            } else {
-                                res.json({success: true, message: "Wifi coupon activated", coupon:outputUser.wifi});
-                            }
-                        });
+                        if (!outputUser){
+                            res.json({success: false, message: "No such user exists"});
+                        } else {
+                            // Removing that coupon from the coupons collection because that coupon is no longer availableto other users
+                            Coupon.findOneAndRemove({_id: coupon._id}).exec(function(err){
+                                if (err){
+                                    console.log(err);
+                                    res.json({success: false, message: "Error occurred while activating the coupon"});
+                                } else {
+                                    res.json({success: true, message: "Wifi coupon activated", coupon:outputUser.wifi});
+                                }
+                            });
+                        }
                     }
-                }
-            });
+                });
+            }
         });
     });
